Add route for camp by state. Adjust map to use state list. Change campsite model to double from decimal.

+++ models/Campsite.js
@@ -19,10 +19,10 @@ Campsite.init(
             type: DataTypes.STRING,
         },
         lat: {
-            type: DataTypes.DECIMAL,
+            type: DataTypes.DOUBLE,
         },
         lng: {
-            type: DataTypes.DECIMAL,
+            type: DataTypes.DOUBLE,
         },
         latlng: {
             type: DataTypes.STRING,
